requirements.txt
pandas
numpy
matplotlib
torch
torchvision
tensorflow
streamlit
scikit-learn
transformers
tf-keras
datasets
accelerate==0.26.0
ipywidgets
requests
gradio
fastapi
uvicorn
speechrecognition
pyttsx3
gspread
google-auth
python-dotenv
